Medical Billing Chatbot/requirements.txt
# Core LLM and RAG libraries
langchain
langchain_core
langchain_community
langchain_experimental
langchain_groq
langchain_huggingface

# Vector stores and embeddings
faiss-cpu
sentence-transformers
chromadb

# Text processing
tiktoken
nltk
spacy
rank_bm25

# Web frameworks
fastapi
uvicorn
pydantic

# Data handling
datasets
pandas
numpy
pyarrow

# Document loaders
unstructured
pypdf
python-docx
pptx

# External integrations
kagglehub[pandas-datasets]
huggingface_hub

# Visualization and metrics
matplotlib
scikit-learn
evaluate

# Utilities
tqdm
python-dotenv
streamlit